move map layout to a new file
