+++ src/managers/TerrainManager.ts
@@ -5,25 +5,6 @@ import { TextureKey, TextureManager } from "./TextureManager";
 import { Terrain, TerrainTileType } from "../models/Terrain";
 import { MapTileKey } from "../models/MapKey";
 
-// prettier-ignore
-export const startingMapArray: MapTileKey[][] = [
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-H", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "_-`", "_-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_W`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_W`", "_W`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-H", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_W`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
-];
-
 export class TerrainManager {
   private static mapKeyIndex = 0;
   private static sizeOfSprite = 16;
