+++ src/mapLayout.ts
@@ -1,0 +1,20 @@
+import { MapTileKey } from "./models/MapKey";
+
+// prettier-ignore
+export const startingMapArray: MapTileKey[][] = [
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-H", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "0-`", "0-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "0-`", "0-`", "0-`", "_-`", "_-`", "0-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_W`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_W`", "_W`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-H", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_W`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+  ["_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`", "_-`"],
+];
